Standardize diagnostics
