+++ BicepNet.Core/BicepWrapper.Restore.cs
@@ -2,7 +2,6 @@ using Bicep.Core.Diagnostics;
 using Bicep.Core.FileSystem;
 using Bicep.Core.Registry;
 using Bicep.Core.Syntax;
-using Bicep.Core.Text;
 using Bicep.Core.Workspaces;
 using Microsoft.Extensions.Logging;
 using System;
@@ -72,7 +71,7 @@ public partial class BicepWrapper
         }
     }
 
-    private static IReadOnlyDictionary<BicepFile, IEnumerable<IDiagnostic>> GetModuleRestoreDiagnosticsByBicepFile(SourceFileGrouping sourceFileGrouping, ImmutableHashSet<ModuleDeclarationSyntax> originalModulesToRestore)
+    private static ImmutableDictionary<BicepFile, ImmutableArray<IDiagnostic>> GetModuleRestoreDiagnosticsByBicepFile(SourceFileGrouping sourceFileGrouping, ImmutableHashSet<ModuleDeclarationSyntax> originalModulesToRestore)
     {
         static IEnumerable<IDiagnostic> GetModuleDiagnosticsPerFile(SourceFileGrouping grouping, BicepFile bicepFile, ImmutableHashSet<ModuleDeclarationSyntax> originalModulesToRestore)
         {
@@ -93,48 +92,6 @@ public partial class BicepWrapper
 
         return sourceFileGrouping.SourceFiles
             .OfType<BicepFile>()
-            .ToDictionary(bicepFile => bicepFile, bicepFile => GetModuleDiagnosticsPerFile(sourceFileGrouping, bicepFile, originalModulesToRestore));
-    }
-
-    private static void LogDiagnostic(Uri fileUri, IDiagnostic diagnostic, ImmutableArray<int> lineStarts)
-    {
-        (int line, int character) = TextCoordinateConverter.GetPosition(lineStarts, diagnostic.Span.Position);
-
-        // build a a code description link if the Uri is assigned
-        var codeDescription = diagnostic.Uri == null ? string.Empty : $" [{diagnostic.Uri.AbsoluteUri}]";
-
-        var message = $"{fileUri.LocalPath}({line + 1},{character + 1}) : {diagnostic.Level} {diagnostic.Code}: {diagnostic.Message}{codeDescription}";
-
-        switch (diagnostic.Level)
-        {
-            case DiagnosticLevel.Off:
-                break;
-            case DiagnosticLevel.Info:
-                logger?.LogInformation(message);
-                break;
-            case DiagnosticLevel.Warning:
-                logger?.LogWarning(message);
-                break;
-            case DiagnosticLevel.Error:
-                logger?.LogError(message);
-                break;
-            default:
-                break;
-        }
-
-        // Increment counters
-        if (diagnostic.Level == DiagnosticLevel.Warning) { WarningCount++; }
-        if (diagnostic.Level == DiagnosticLevel.Error) { ErrorCount++; }
-    }
-
-    private static void LogDiagnostics(IReadOnlyDictionary<BicepFile, IEnumerable<IDiagnostic>> diagnosticsByBicepFile)
-    {
-        foreach (var (bicepFile, diagnostics) in diagnosticsByBicepFile)
-        {
-            foreach (var diagnostic in diagnostics)
-            {
-                LogDiagnostic(bicepFile.FileUri, diagnostic, bicepFile.LineStarts);
-            }
-        }
+            .ToImmutableDictionary(bicepFile => bicepFile, bicepFile => GetModuleDiagnosticsPerFile(sourceFileGrouping, bicepFile, originalModulesToRestore).ToImmutableArray());
     }
 }
